poc330回合主干

Created-by: lilongbin 00657083
Author-id: 534576
MR-id: 26349323
Commit-by: l00657083;zhousipei;huanglong 30059694;huangyuyang 00619579;dengzhaochu 00807371;lilongbin 00657083;houyongfei;hyy_cyan;x00613198;panmingyi;huanglong;kongxinghan 00624780;x30027624;w00630100;lwx1336275;k00624780;liujingxiang;d00807371;l30069941 ;liangruifeng 30069941;s30031572
Merged-by: huangyuyang 00619579
E2E-issues: AR20250217767806
Description: 全量+社区用例验证已OK
![image](/uploads/bf46f388-c2f2-4de7-b8e1-99b19ad75fb6/1743990813993.png '1743990813993.png')

See merge request: Kunpeng-Computing-DC/BigData/OmniOperatorJIT!970

+++ core/test/operator/aggregation_with_expr_test.cpp
@@ -1056,11 +1056,11 @@ TEST(HashAggregationWithExprOperatorTest, test_hashagg_spill_with_invalid_config
     int64_t data2[] = {5L, 3L, 2L, 6L, 1L, 4L, 7L, 8L};
 
     const int32_t groupByNum = 1;
-    DataTypes sourceTypes(std::vector<DataTypePtr>({ LongType(), LongType() }));
+    DataTypes sourceTypes(std::vector<DataTypePtr>({DoubleType(), LongType()}));
     DataTypes aggOutputTypes(std::vector<DataTypePtr>({ LongType(), LongType() }));
 
     // groupByKeys
-    std::vector<Expr *> groupByKeys = { new FieldExpr(0, LongType()) };
+    std::vector<Expr *> groupByKeys = {new FieldExpr(0, DoubleType())};
 
     // aggKeys
     std::vector<Expr *> aggKeys1 = { new FieldExpr(1, LongType()) };
